requirements.txt
streamlit>=1.30,<2.0
pandas>=2.0,<3.0
pymodbus==3.6.9
plotly>=5.18,<6.0
openpyxl>=3.1,<4.0
python-dotenv>=1.0,<2.0
reportlab>=4.0,<5.0
fastapi>=0.100,<1.0
uvicorn>=0.20,<1.0
prometheus_client>=0.20,<1.0
scikit-learn>=1.3.0,<2.0
websockets>=11.0,<13.0
paho-mqtt>=1.6,<3.0